Extract FeatureCompiler
