+++ core/src/main/java/cucumber/runtime/FeatureCompiler.java
@@ -1,0 +1,21 @@
+package cucumber.runtime;
+
+import cucumber.runtime.model.CucumberFeature;
+import gherkin.events.PickleEvent;
+import gherkin.pickles.Compiler;
+import gherkin.pickles.Pickle;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public final class FeatureCompiler {
+    private final Compiler compiler = new Compiler();
+
+    public List<PickleEvent> compileFeature(CucumberFeature feature) {
+        List<PickleEvent> pickleEvents = new ArrayList<PickleEvent>();
+        for (Pickle pickle : compiler.compile(feature.getGherkinFeature())) {
+            pickleEvents.add(new PickleEvent(feature.getUri(), pickle));
+        }
+        return pickleEvents;
+    }
+}
